decoupled the backend app.py into 3 seperate controllers depending on the topic
the purpose is to provide clarity regarding the stream of data within the backend as it is flowing towards the frontend and the database.

the controllers are categorized by their respective topics: route, plow, driver
